+++ examples/low_level_api/low_level_api_llama_cpp.py
@@ -11,20 +11,34 @@ MODEL_PATH = os.environ.get('MODEL', "../models/7B/ggml-model.bin")
 
 prompt = b"\n\n### Instruction:\nWhat is the capital of France?\n\n### Response:\n"
 
-lparams = llama_cpp.llama_context_default_params()
+lparams = llama_cpp.llama_model_default_params()
+cparams = llama_cpp.llama_context_default_params()
 model = llama_cpp.llama_load_model_from_file(MODEL_PATH.encode('utf-8'), lparams)
-ctx = llama_cpp.llama_new_context_with_model(model, lparams)
+ctx = llama_cpp.llama_new_context_with_model(model, cparams)
 
 # determine the required inference memory per token:
 tmp = [0, 1, 2, 3]
-llama_cpp.llama_eval(ctx, (llama_cpp.c_int * len(tmp))(*tmp), len(tmp), 0, N_THREADS)
+llama_cpp.llama_eval(
+            ctx = ctx, 
+            tokens=(llama_cpp.c_int * len(tmp))(*tmp),
+            n_tokens=len(tmp),
+            n_past=0
+        )# Deprecated
 
 n_past = 0
 
 prompt = b" " + prompt
 
 embd_inp = (llama_cpp.llama_token * (len(prompt) + 1))()
-n_of_tok = llama_cpp.llama_tokenize(ctx, prompt, embd_inp, len(embd_inp), True)
+n_of_tok = llama_cpp.llama_tokenize(
+    model=model,
+    text=bytes(str(prompt),'utf-8'),
+    text_len=len(embd_inp), 
+    tokens=embd_inp,
+    n_max_tokens=len(embd_inp),
+    add_bos=False,
+    special=False
+)
 embd_inp = embd_inp[:n_of_tok]
 
 n_ctx = llama_cpp.llama_n_ctx(ctx)
@@ -49,8 +63,11 @@ presence_penalty = 0.0
 while remaining_tokens > 0:
     if len(embd) > 0:
         llama_cpp.llama_eval(
-            ctx, (llama_cpp.c_int * len(embd))(*embd), len(embd), n_past, N_THREADS
-        )
+            ctx = ctx, 
+            tokens=(llama_cpp.c_int * len(embd))(*embd),
+            n_tokens=len(embd),
+            n_past=n_past
+        )# Deprecated
 
     n_past += len(embd)
     embd = []
@@ -93,7 +110,7 @@ while remaining_tokens > 0:
         for id in embd:
             size = 32
             buffer = (ctypes.c_char * size)()
-            n = llama_cpp.llama_token_to_piece_with_model(
+            n = llama_cpp.llama_token_to_piece(
                 model, llama_cpp.llama_token(id), buffer, size)
             assert n <= size
             print(