+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.19]
+
+- Update llama.cpp to ggerganov/llama.cpp@0b871f1a04ef60e114bbe43004fd9c21114e802d
+- Fix #569: stop parameter in chat completion api should accept str by @abetlen in 128dc4731fa846ead7e684a137ca57d8931b8899
+- Document server host and port parameters by @jamesbraza in #768
+- Do not set grammar to None when initializing LlamaGrammar by @mthuurne in #834
+- Add mistrallite, intel, and openchat formats by @fakerybakery in #927
+- Add support for min_p parameter by @tk-master in #921
+- Fix #929: tokenizer adding leading space when generating from empty prompt by @abetlen in a34d48014192771d2e308a76c22f33bc0318d983
+- Fix low level api example by @zocainViken in #925
+- Fix missing package in openblas docker image by @ZisisTsatsas in #920
+
 ## [0.2.18]
 
 - Update llama.cpp to ggerganov/llama.cpp@6bb4908a17150b49373b5f977685b2e180a04f6f